requirements.txt
beautifulsoup4==4.4.1               # MIT
django-webtest==1.7.8               # MIT
Django==1.9.1                       # 3-clause BSD
psycopg2==2.6.1                     # GNU LGPL
six==1.10.0                         # MIT
waitress==0.8.10                    # ZPL
WebOb==1.5.1                        # MIT
WebTest==2.0.20                     # MIT
wheel==0.24.0                       # MIT
django-taggit==0.18.0               # BSD
django-compressor==2.0              # MIT
django-libsass==0.6.0               # BSD
python-dateutil==2.5.0              # Simplified BSD
